update nesfab, fix bugs
nesfab hotfix seems to fix all of the problems in the previous commit.
two other changes: first, the hotfix uses a new ppu_scroll routine that
takes more time (but is safer), so the old one was copied in to be used.
second, wasting time during the midframe scroll is refactored to use a
dummy hardware write, which generates much more consistent code. it has
a granularity of 4 cpu cycles