begin major refactor of code structure
this is the first of several changes made to facilitate submodes. the
two concrete things done here are using bitfields for similar
operations, and moving the update_state logic outside of the recording
clause (along with a rename).

more abstractly, a proper implementation of submodes requires a
different way of thinking about state. We now differentiate between the
*core state* which is immediately updated when polls are processed, and
display state which uses the core state to determine what needs to be
displayed. this way, all submodes can rely on the same core state, and
use their own display implementations to customize its representation.